requirements.txt
langchain
langchain-openai
langgraph
langchain-community
pydantic
python-dotenv
requests